+++ src/Admin/Assets.php
@@ -46,17 +46,17 @@ final class Assets {
 	private static function inlineConfig(): string {
 		$d  = '';
 		$d .= '( function ( global ) {';
-		$d .=   '"use strict";';
-		$d .=   'const tmq = global.tmq = global.tmq || {};';
-		$d .=   'tmq.restUrl = "' . esc_url( wp_make_link_relative( rest_url() ) ) . '";';
-		$d .=   'tmq.restNonce = "' . esc_html( wp_create_nonce( 'wp_rest' ) ) . '";';
-		$d .=   'tmq.i18n = tmq.i18n || {};';
-		$d .=   'tmq.i18n.errorLoadingMessage = "' . esc_js( __( 'There was an error loading the message.', 'total-mail-queue' ) ) . '";';
-		$d .=   'tmq.i18n.confirmDelete = "' . esc_js( __( 'Are you sure you want to delete the selected items? This action cannot be undone.', 'total-mail-queue' ) ) . '";';
-		$d .=   'tmq.i18n.testing = "' . esc_js( __( 'Testing...', 'total-mail-queue' ) ) . '";';
-		$d .=   'tmq.i18n.testConnection = "' . esc_js( __( 'Test Connection', 'total-mail-queue' ) ) . '";';
-		$d .=   'tmq.ajaxUrl = "' . esc_url( admin_url( 'admin-ajax.php' ) ) . '";';
-		$d .=   'tmq.testSmtpNonce = "' . esc_js( wp_create_nonce( ConnectionTester::NONCE ) ) . '";';
+		$d .= '"use strict";';
+		$d .= 'const tmq = global.tmq = global.tmq || {};';
+		$d .= 'tmq.restUrl = "' . esc_url( wp_make_link_relative( rest_url() ) ) . '";';
+		$d .= 'tmq.restNonce = "' . esc_html( wp_create_nonce( 'wp_rest' ) ) . '";';
+		$d .= 'tmq.i18n = tmq.i18n || {};';
+		$d .= 'tmq.i18n.errorLoadingMessage = "' . esc_js( __( 'There was an error loading the message.', 'total-mail-queue' ) ) . '";';
+		$d .= 'tmq.i18n.confirmDelete = "' . esc_js( __( 'Are you sure you want to delete the selected items? This action cannot be undone.', 'total-mail-queue' ) ) . '";';
+		$d .= 'tmq.i18n.testing = "' . esc_js( __( 'Testing...', 'total-mail-queue' ) ) . '";';
+		$d .= 'tmq.i18n.testConnection = "' . esc_js( __( 'Test Connection', 'total-mail-queue' ) ) . '";';
+		$d .= 'tmq.ajaxUrl = "' . esc_url( admin_url( 'admin-ajax.php' ) ) . '";';
+		$d .= 'tmq.testSmtpNonce = "' . esc_js( wp_create_nonce( ConnectionTester::NONCE ) ) . '";';
 		$d .= '}) ( this );';
 		return $d;
 	}